move libs

+++ src/build.bat
@@ -7,22 +7,26 @@ ctime -begin .\logs\text_editor.ctm
 
 del *.pdb > NUL 2> NUL
 
-cp ..\src\libs\SDL2\lib\x64\SDL2.dll SDL2.dll
-cp ..\src\libs\SDL2_ttf\lib\x64\SDL2_ttf.dll SDL2_ttf.dll
-cp ..\src\libs\SDL2_ttf\lib\x64\libfreetype-6.dll libfreetype-6.dll
-cp ..\src\libs\SDL2_ttf\lib\x64\zlib1.dll zlib1.dll
+IF NOT EXIST SDL2.dll   cp ..\src\libs\SDL2\lib\x64\SDL2.dll SDL2.dll
+IF NOT EXIST SDL2_ttf.dll   cp ..\src\libs\SDL2_ttf\lib\x64\SDL2_ttf.dll SDL2_ttf.dll
+IF NOT EXIST libfreetype-6.dll   cp ..\src\libs\SDL2_ttf\lib\x64\libfreetype-6.dll libfreetype-6.dll
+IF NOT EXIST zlib1.dll   cp ..\src\libs\SDL2_ttf\lib\x64\zlib1.dll zlib1.dll
+
+:: shared libs folder path
+set SharedLibsPath=D:\_shared_libs
 
 :: -O2 optimization level 2
 :: -Od for debbugging, no optimization
 ::  -WX -W4    I should be using those
+:: -nologo   does not show the compiler, architecture (x64) and the output files (exe, libs, ...), ...
 set CommonCompilerFlags=-diagnostics:column -diagnostics:caret -Od -WL -nologo -fp:fast -fp:except- -Gm- -GR- -EHa- -Zo -Oi -WX -W4 -wd4201 -wd4100 -wd4189 -wd4505 -wd4127 -FC -Z7 /EHsc
 :: /EHsc because of SDL i guess
 
 :: Additional Include Directories
-set AdditionalIncludeDirectories=/I D:\tebtro\iml\src /I..\src\libs\SDL2\include /I ..\src\libs\SDL2_ttf\include
+set AdditionalIncludeDirectories=/I D:\tebtro\iml\src /I..\libs\SDL2\include /I ..\libs\SDL2_ttf\include
 
 :: Additional Linker Options
-set AdditionalLinkerFlags=-incremental:no -opt:ref /SUBSYSTEM:CONSOLE /LIBPATH:..\src\libs\SDL2\lib\x64 SDL2.lib SDL2main.lib SDL2test.lib /LIBPATH:..\src\libs\SDL2_ttf\lib\x64 SDL2_ttf.lib
+set AdditionalLinkerFlags=-incremental:no -opt:ref /SUBSYSTEM:CONSOLE /LIBPATH:..\libs\SDL2\lib\x64 SDL2.lib SDL2main.lib SDL2test.lib /LIBPATH:..\libs\SDL2_ttf\lib\x64 SDL2_ttf.lib
 
 :: Libraries
 set Libraries=user32.lib gdi32.lib kernel32.lib glu32.lib gdi32.lib  opengl32.lib
